Do not cache topic in proxy endpoint (#207)
The only place it was used for more than printing the name (which is
available via the QoS object) was in deserializing the sample, but there
is never a need to deserialize a sample if there is no local reader.  So
instead of caching the topic object in the proxy endpoint, take it from
a reader to which it is to be delivered and avoid having to keep things
in sync.

Signed-off-by: Erik Boasson <eb@ilities.com>