Deliver initial samples to all readers
While commit 894e47c7b33001fc7741b74cb08298d2582172db addressed most of
the issues with a new reader possibly not accepting all samples a proxy
writer published after it matched that reader, it failed to do so
properly for 2nd and later readers.  The cause was, once again,
prematurely considering the reader to be "in sync" with the proxy
writer and not requesting any old samples.

The desire to consider them "in sync" from the very beginning is that it
avoids a lot of network traffic.  With this commit, all new readers are
treated equally, despite the additional traffic that generates.

Signed-off-by: Erik Boasson <eb@ilities.com>